Split FdataBasis in another file
